+++ bin/install.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-# DATE: 2020-01-11
-# VERSION: 1.9.2
+# DATE: 2020-04-19
+# VERSION: 1.10.0
 
 ###############################################################################
 # Set variables                                                               #
@@ -29,6 +29,16 @@ while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
 # Install Homebrew apps
 printf "📦 Installing Apps...\n"
 brew bundle --file=$SETUP/Brewfile
+
+###############################################################################
+# Install Conda                                                               #
+###############################################################################
+
+wget https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O ~/miniconda.sh
+zsh ~/miniconda.sh -b -p $HOME/miniconda
+source $HOME/miniconda/bin/activate
+conda init zsh
+conda deactivate
 
 ###############################################################################
 # Final touches                                                               #